+++ base/floatfuncs.jl
@@ -132,7 +132,7 @@ function _signif_og(x, digits, base)
 end
 
 function signif(x::Real, digits::Integer, base::Integer=10)
-    digits < 1 && throw(DomainError())
+    digits < 1 && throw(DomainError(digits, "`digits` cannot be less than 1."))
 
     x = float(x)
     (x == 0 || !isfinite(x)) && return x